# docs/swagger.yaml
basePath: /v1
definitions:
...
  entity.Video:
    properties:
      captionAvailability:
        $ref: '#/definitions/entity.VideoCaptionAvailability'
      channel:
        $ref: '#/definitions/entity.VideoChannel'
      createdAt:
        type: string
      durationSeconds:
        example: 300
        type: integer
      id:
        example: 1
        type: integer
      language:
        $ref: '#/definitions/entity.VideoLanguage'
      level:
        $ref: '#/definitions/entity.VideoLevel'
      status:
        example: published
        type: string
      thumbnailUrl:
        type: string
      title:
        example: Learn English with Music
        type: string
      topics:
        items:
          $ref: '#/definitions/entity.VideoTopic'
        type: array
      updatedAt:
        type: string
      videoUrl:
        example: https://www.youtube.com/watch?v=dQw4w9WgXcQ
        type: string
      youtubeId:
        example: dQw4w9WgXcQ
        type: string
    type: object
  entity.VideoCaptionAvailability:
    properties:
      captionCount:
        example: 120
        type: integer
      hasOriginal:
        example: true
        type: boolean
      translations:
        items:
          $ref: '#/definitions/entity.VideoCaptionTranslationAvailability'
        type: array
    type: object
  entity.VideoCaptionTranslationAvailability:
    properties:
      isComplete:
        example: true
        type: boolean
      languageCode:
        example: vi
        type: string
      languageId:
        example: 2
        type: integer
      languageName:
        example: Vietnamese
        type: string
      translatedCaptionCount:
        example: 120
        type: integer
    type: object
  entity.VideoChannel:
    properties:
      avatarUrl:
        type: string
      channelYoutubeId:
        type: string
      id:
        example: 3
        type: integer
      name:
        example: Google Developers
        type: string
    type: object
  entity.VideoLanguage:
    properties:
      code:
        example: en
        type: string
      flagEmoji:
        example: "\U0001F1EC\U0001F1E7"
        type: string
      id:
        example: 1
        type: integer
      name:
        example: English
        type: string
    type: object
  entity.VideoLevel:
    properties:
      code:
        example: A2
        type: string
      id:
        example: 2
        type: integer
      name:
        example: Elementary
        type: string
    type: object
...
  entity.VideoTopic:
    properties:
      iconUrl:
        type: string
      id:
        example: 5
        type: integer
      name:
        example: Music
        type: string
      slug:
        example: music
        type: string
    type: object
...
  github_com_evrone_go-clean-template_internal_controller_restapi_v1_request.ImportYouTubeVideo:
    properties:
      audioLanguageCode:
        example: en
        maxLength: 35
        type: string
      youtubeUrl:
        example: https://www.youtube.com/watch?v=dQw4w9WgXcQ
        maxLength: 2048
        type: string
    required:
    - youtubeUrl
    type: object
...
  github_com_evrone_go-clean-template_internal_controller_restapi_v1_request.YouTubePreview:
    properties:
      youtubeUrl:
        example: https://www.youtube.com/watch?v=dQw4w9WgXcQ
        maxLength: 2048
        type: string
    required:
    - youtubeUrl
    type: object
  github_com_evrone_go-clean-template_internal_controller_restapi_v1_response.OriginalVideoCaptions:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_response.VideoCaptionItem'
        type: array
      languageCode:
        type: string
      languageId:
        type: integer
      total:
        type: integer
      video_state:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.VideoState'
      videoId:
        type: integer
    type: object
  github_com_evrone_go-clean-template_internal_controller_restapi_v1_response.VideoCaptionItem:
    properties:
      dictation_completed:
        example: true
        type: boolean
      endTimeMs:
        example: 3500
        type: integer
      id:
        example: 1
        type: integer
      sentenceOrder:
        example: 1
        type: integer
      startTimeMs:
        example: 1000
        type: integer
      text:
        example: Hello
        type: string
    type: object
...
  github_com_evrone_go-clean-template_internal_entity.FeatureKey:
    enum:
    - youtube_import
    - shadowing_assessment
    type: string
    x-enum-varnames:
    - FeatureYouTubeImport
    - FeatureShadowingAssessment
...
  github_com_evrone_go-clean-template_internal_entity.ShadowingAttempt:
    properties:
      accuracy_score:
        example: 79
        type: number
      caption_id:
        example: 10
        type: integer
      completeness_score:
        example: 100
        type: number
      created_at:
        type: string
      duration_ms:
        example: 2300
        type: integer
      fluency_score:
        example: 86
        type: number
      id:
        example: 123
        type: integer
      locale:
        example: en-US
        type: string
      passed:
        example: true
        type: boolean
      pronunciation_score:
        example: 82
        type: number
      prosody_score:
        example: 75
        type: number
      provider:
        example: azure
        type: string
      reference_text:
        example: Good morning.
        type: string
      video_id:
        example: 1
        type: integer
      words:
        items:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.ShadowingWordResult'
        type: array
    type: object
  github_com_evrone_go-clean-template_internal_entity.ShadowingAttemptList:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.ShadowingAttempt'
        type: array
      total:
        example: 3
        type: integer
    type: object
  github_com_evrone_go-clean-template_internal_entity.ShadowingWordResult:
    properties:
      accuracy_score:
        example: 54
        type: number
      error_type:
        example: Mispronunciation
        type: string
      order:
        example: 0
        type: integer
      phonemes:
        type: object
      word:
        example: comfortable
        type: string
    type: object
...
  github_com_evrone_go-clean-template_internal_entity.UserYouTubeVideoResult:
    properties:
      captionImported:
        example: true
        type: boolean
      captionSource:
        example: groq_whisper
        type: string
      reused:
        example: false
        type: boolean
      video:
        $ref: '#/definitions/entity.Video'
    type: object
...
  github_com_evrone_go-clean-template_internal_entity.VideoCaptionItem:
    properties:
      endTimeMs:
        type: integer
      id:
        type: integer
      sentenceOrder:
        type: integer
      startTimeMs:
        type: integer
      text:
        type: string
    type: object
  github_com_evrone_go-clean-template_internal_entity.VideoCaptions:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.VideoCaptionItem'
        type: array
      languageCode:
        type: string
      languageId:
        type: integer
      total:
        type: integer
      videoId:
        type: integer
    type: object
...
  github_com_evrone_go-clean-template_internal_entity.VideoState:
    properties:
      last_position_seconds:
        example: 75
        type: integer
      saved:
        example: false
        type: boolean
      watched:
        example: true
        type: boolean
    type: object
...
  github_com_evrone_go-clean-template_internal_entity.YouTubeVideoPreview:
    properties:
      channelName:
        type: string
      channelYoutubeId:
        type: string
      durationSeconds:
        example: 300
        type: integer
      thumbnailUrl:
        type: string
      title:
        example: Video title
        type: string
      youtubeId:
        example: dQw4w9WgXcQ
        type: string
    type: object
  internal_controller_restapi_v1.quotaErrorDoc:
    properties:
      code:
        type: string
      error:
        type: string
      feature:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.FeatureKey'
      limit:
        type: integer
      remaining:
        type: integer
      resetAt:
        type: string
      upgradeRequired:
        type: boolean
      used:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: User API with Firebase authentication and role-based access control
  title: Go Clean Template API
  version: "1.0"
paths:
...
  /videos/{videoId}/captions:
    get:
      description: Returns original captions immediately without requesting a translation
      operationId: user-video-original-captions
      parameters:
      - description: Video ID
        in: path
        name: videoId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_response.OriginalVideoCaptions'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get original video captions
      tags:
      - user-videos
  /videos/{videoId}/captions/translation:
    get:
      description: Returns captions translated to the user's native language; missing
        translations are generated and persisted before returning
      operationId: user-video-translated-captions
      parameters:
      - description: Video ID
        in: path
        name: videoId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.VideoCaptions'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1.quotaErrorDoc'
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get translated video captions
      tags:
      - user-videos
...
  /videos/{videoId}/shadowing-attempts:
    get:
      description: Return the current user's attempts for a video, optionally filtered
        by caption.
      operationId: list-shadowing-attempts
      parameters:
      - description: Video ID
        in: path
        name: videoId
        required: true
        type: integer
      - description: Caption ID
        in: query
        name: caption_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.ShadowingAttemptList'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List shadowing attempts
      tags:
      - shadowing
  /videos/{videoId}/shadowing-attempts/{captionId}:
    post:
      consumes:
      - multipart/form-data
      description: Upload WAV PCM 16 kHz or OGG/Opus audio and assess it against the
        selected caption.
      operationId: assess-shadowing
      parameters:
      - description: Video ID
        in: path
        name: videoId
        required: true
        type: integer
      - description: Caption ID
        in: path
        name: captionId
        required: true
        type: integer
      - description: Recorded audio (max 5 MB)
        in: formData
        name: audio
        required: true
        type: file
      - description: Azure BCP-47 locale; defaults from video language
        in: formData
        name: locale
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.ShadowingAttempt'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "413":
          description: Request Entity Too Large
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1.quotaErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Assess a shadowing attempt
      tags:
      - shadowing
  /videos/import-youtube:
    post:
      consumes:
      - application/json
      description: Reuses an existing video and captions when available; otherwise
        downloads normalized audio and transcribes it with Groq Whisper. audioLanguageCode
        is an optional spoken-language hint.
      operationId: user-import-youtube-video
      parameters:
      - description: YouTube video import
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_request.ImportYouTubeVideo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.UserYouTubeVideoResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add a YouTube video for the current user
      tags:
      - user-videos
...
  /videos/youtube-preview:
    post:
      consumes:
      - application/json
      description: Returns YouTube metadata without saving the video or downloading
        audio
      operationId: user-preview-youtube-video
      parameters:
      - description: YouTube URL or ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_request.YouTubePreview'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.YouTubeVideoPreview'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Preview a YouTube video
      tags:
      - user-videos
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"